+++ src/Components/Pages/TransferPage.tsx
@@ -8,6 +8,7 @@ import {
   Typography,
   // QuestionCircleSvg,
   SelectInput,
+  NavLink,
 } from "@chainsafe/common-components";
 import { Form, Formik } from "formik";
 import AddressInput from "../Custom/AddressInput";
@@ -23,11 +24,12 @@ import { useNetworkManager } from "../../Contexts/NetworkManagerContext";
 import NetworkUnsupportedModal from "../../Modules/NetworkUnsupportedModal";
 import { isValidSubstrateAddress } from "../../Utils/Helpers";
 import { useHomeBridge } from "../../Contexts/HomeBridgeContext";
-import ETHIcon from "../../media/tokens/eth.svg";
+import { ReactComponent as ETHIcon } from "../../media/tokens/eth.svg";
 import WETHIcon from "../../media/tokens/weth.svg";
 import DAIIcon from "../../media/tokens/dai.svg";
 import celoUSD from "../../media/tokens/cusd.svg";
-import CEREIcon from "../../media/tokens/cere-token.svg";
+import { ReactComponent as CEREIcon } from "../../media/tokens/cere-token.svg";
+import styles from "../../Constants/constants";
 
 const PredefinedIcons: any = {
   ETHIcon: ETHIcon,
@@ -40,11 +42,22 @@ const PredefinedIcons: any = {
 const showImageUrl = (url?: string) =>
   url && PredefinedIcons[url] ? PredefinedIcons[url] : url;
 
+const borderColor = "#C4C4C4";
+
 const useStyles = makeStyles(({ constants, palette }: ITheme) =>
   createStyles({
     root: {
-      padding: constants.generalUnit * 6,
+      fontFamily: styles.primaryFont,
+      padding: constants.generalUnit * 3,
       position: "relative",
+      backgroundColor: "white",
+    },
+    title: {
+      fontSize: 24,
+      fontWeight: "bold",
+      color: "black",
+      textAlign: "center",
+      paddingBottom: constants.generalUnit * 3.75,
     },
     walletArea: {
       display: "flex",
@@ -52,9 +65,15 @@ const useStyles = makeStyles(({ constants, palette }: ITheme) =>
       alignItems: "center",
       justifyContent: "center",
       width: "100%",
+      marginBottom: constants.generalUnit * 3,
     },
     connectButton: {
-      margin: `${constants.generalUnit * 3}px 0 ${constants.generalUnit * 6}px`,
+      background: "linear-gradient(105.79deg, #A700E1 1.84%, #0024E2 102.94%)",
+      fontWeight: "bold",
+      borderRadius: 5,
+      "&:hover": {
+        color: "white",
+      },
     },
     connecting: {
       textAlign: "center",
@@ -72,16 +91,16 @@ const useStyles = makeStyles(({ constants, palette }: ITheme) =>
     },
     changeButton: {
       cursor: "pointer",
+      color: styles.primaryTextColor,
+      fontSize: 11,
+      fontFamily: `${styles.secondaryFont} !important`,
     },
     networkName: {
-      padding: `${constants.generalUnit * 2}px ${
-        constants.generalUnit * 1.5
-      }px`,
-      border: `1px solid ${palette.additional["gray"][6]}`,
+      padding: `${constants.generalUnit / 1.5}px ${constants.generalUnit}px`,
+      border: `1px solid ${borderColor}`,
       borderRadius: 2,
       color: palette.additional["gray"][9],
       marginTop: constants.generalUnit,
-      marginBottom: constants.generalUnit * 3,
     },
     formArea: {
       "&.disabled": {
@@ -112,6 +131,11 @@ const useStyles = makeStyles(({ constants, palette }: ITheme) =>
           borderBottomRightRadius: 0,
           borderTopRightRadius: 0,
           borderRight: 0,
+          border: `1px solid ${borderColor}`,
+          "&:hover": {
+            border: `1px solid ${borderColor}`,
+            borderRight: 0,
+          },
         },
       },
       "& span:last-child.error": {
@@ -124,16 +148,15 @@ const useStyles = makeStyles(({ constants, palette }: ITheme) =>
       borderBottomLeftRadius: 0,
       borderTopLeftRadius: 0,
       left: -1,
-      color: palette.additional["gray"][8],
-      backgroundColor: palette.additional["gray"][3],
-      borderColor: palette.additional["gray"][6],
+      color: styles.primaryTextColor,
+      borderLeft: 0,
+      backgroundColor: "white",
+      borderColor: `${borderColor} !important`,
       "&:hover": {
-        borderColor: palette.additional["gray"][6],
-        backgroundColor: palette.additional["gray"][7],
-        color: palette.common.white.main,
-      },
-      "&:focus": {
-        borderColor: palette.additional["gray"][6],
+        borderColor: `${borderColor} !important`,
+        backgroundColor: "white",
+        color: styles.primaryTextColor,
+        borderLeft: 0,
       },
     },
     currencySelector: {
@@ -147,8 +170,13 @@ const useStyles = makeStyles(({ constants, palette }: ITheme) =>
     address: {
       margin: 0,
       marginBottom: constants.generalUnit * 3,
+      fontSize: 8,
     },
-    addressInput: {},
+    addressInput: {
+      "& > div > input": {
+        fontSize: "12px !important",
+      },
+    },
     generalInput: {
       "& > span": {
         marginBottom: constants.generalUnit,
@@ -184,10 +212,11 @@ const useStyles = makeStyles(({ constants, palette }: ITheme) =>
       flexWrap: "wrap",
       justifyContent: "space-between",
       marginBottom: constants.generalUnit,
+      fontFamily: styles.secondaryFont,
+      color: styles.greyColor,
       "& > *": {
         display: "block",
         width: "50%",
-        color: palette.additional["gray"][8],
         marginBottom: constants.generalUnit / 2,
         "&:nth-child(even)": {
           textAlign: "right",
@@ -195,7 +224,46 @@ const useStyles = makeStyles(({ constants, palette }: ITheme) =>
       },
     },
     accountSelector: {
-      marginBottom: 24,
+      marginBottom: constants.generalUnit * 3,
+    },
+    transferButton: {
+      background: "linear-gradient(105.79deg, #A700E1 1.84%, #0024E2 102.94%)",
+      fontWeight: "bold",
+      borderRadius: 5,
+      "&:hover": {
+        color: "white",
+      },
+    },
+    footer: {
+      color: styles.primaryTextColor,
+      fontSize: 12,
+      display: "flex",
+      justifyContent: "space-between",
+      alignItems: "center",
+      paddingTop: constants.generalUnit * 2,
+      fontFamily: styles.secondaryFont,
+      fontStyle: "normal",
+      fontWeight: "bold",
+    },
+    footerText: {
+      color: styles.primaryTextColor,
+      fontSize: 12,
+      fontStyle: "normal",
+      fontWeight: "bold",
+    },
+    inputBorder: {
+      border: `1px solid ${borderColor}`,
+    },
+    inputLabel: {
+      fontSize: 12,
+      fontWeight: "normal",
+      fontStyle: "normal",
+      color: "black",
+    },
+    networkIcon: {
+      width: constants.generalUnit * 1.5,
+      height: constants.generalUnit * 1.5,
+      marginRight: constants.generalUnit / 2,
     },
   })
 );
@@ -334,6 +402,7 @@ const TransferPage = () => {
 
   return (
     <article className={classes.root}>
+      <div className={classes.title}>Transfer Tokens (ERC20 to Native)</div>
       <div className={classes.walletArea}>
         {!isReady ? (
           <Button
@@ -355,22 +424,24 @@ const TransferPage = () => {
         ) : (
           <section className={classes.connected}>
             <div>
-              <Typography variant="body1">Home network</Typography>
-              <Typography
-                className={classes.changeButton}
-                variant="body1"
-                onClick={() => setChangeNetworkOpen(true)}
-              >
-                Change
-              </Typography>
+              <div className={classes.inputLabel}>Home Network</div>
+              {/* ToDo: uncomment after enabling Cere -> Eth flow */}
+              {/*<Typography*/}
+              {/*  className={classes.changeButton}*/}
+              {/*  variant="body1"*/}
+              {/*  onClick={() => setChangeNetworkOpen(true)}*/}
+              {/*>*/}
+              {/*  Change*/}
+              {/*</Typography>*/}
             </div>
-            <Typography
-              component="h2"
-              variant="h2"
-              className={classes.networkName}
-            >
+            <div className={classes.networkName}>
+              {walletType === "Ethereum" ? (
+                <ETHIcon className={classes.networkIcon} />
+              ) : (
+                <CEREIcon className={classes.networkIcon} />
+              )}
               {homeConfig?.name}
-            </Typography>
+            </div>
           </section>
         )}
       </div>
@@ -474,12 +545,6 @@ const TransferPage = () => {
                       value: t,
                       label: (
                         <div className={classes.tokenItem}>
-                          {tokens[t]?.imageUri && (
-                            <img
-                              src={showImageUrl(tokens[t]?.imageUri)}
-                              alt={tokens[t]?.symbol}
-                            />
-                          )}
                           <span>{tokens[t]?.symbol || t}</span>
                         </div>
                       ),
@@ -493,7 +558,7 @@ const TransferPage = () => {
                 disabled={!destinationChainConfig}
                 name="receiver"
                 label="Destination Address"
-                placeholder="Please enter the receiving address"
+                placeholder="Please enter the receiving address..."
                 className={classes.address}
                 classNames={{
                   input: classes.addressInput,
@@ -516,19 +581,39 @@ const TransferPage = () => {
               }
             />
             <section>
-              <Button type="submit" fullsize variant="primary">
-                Start transfer
+              <Button
+                type="submit"
+                fullsize
+                variant="primary"
+                className={classes.transferButton}
+              >
+                Start Transfer
               </Button>
             </section>
-            {/* <section>
-            <QuestionCircleSvg
-              onClick={() => setAboutOpen(true)}
-              className={classes.faqButton}
-            />
-          </section> */}
           </Form>
         )}
       </Formik>
+      <section className={classes.footer}>
+        <NavLink
+          style={{ textDecoration: "none" }}
+          className={classes.footerText}
+          to={{ pathname: "https://cere.network/" }}
+          target="_blank"
+        >
+          Cere Homepage
+        </NavLink>
+        <NavLink
+          style={{ textDecoration: "none" }}
+          className={classes.footerText}
+          to={{
+            pathname:
+              "https://explorer.cere.network/?rpc=wss%3A%2F%2Frpc.mainnet.cere.network%3A9945#/staking",
+          }}
+          target="_blank"
+        >
+          Cere Staking
+        </NavLink>
+      </section>
       <AboutDrawer open={aboutOpen} close={() => setAboutOpen(false)} />
       <ChangeNetworkDrawer
         open={changeNetworkOpen}