+++ src/Constants/constants.ts
@@ -1,0 +1,8 @@
+const styles = {
+  primaryTextColor: "#5C4DCF",
+  primaryFont: "Sora, sans-serif",
+  secondaryFont: "SF Pro Display, sans-serif",
+  greyColor: "#717171",
+};
+
+export default styles;