+++ src/Contexts/Adaptors/EVMAdaptors/EVMHomeAdaptorProvider.tsx
@@ -19,13 +19,6 @@ import { decodeAddress } from "@polkadot/util-crypto";
 
 import { hasTokenSupplies, getPriceCompatibility } from "./helpers";
 
-const resetAllowanceLogicFor = [
-  "0xdac17f958d2ee523a2206206994597c13d831ec7", //USDT
-  "0x874069Fa1Eb16D44d622F2e0Ca25eeA172369bC1", //cUSD CELO
-  // "0xe09523d86d9b788BCcb580d061605F31FCe69F51", //сTST CELO cUSD on Rinkeby
-  //Add other offending tokens here
-];
-
 export const EVMHomeAdaptorProvider = ({
   children,
 }: IHomeBridgeProviderProps) => {
@@ -45,20 +38,20 @@ export const EVMHomeAdaptorProvider = ({
 
   const getNetworkName = (id: any) => {
     switch (Number(id)) {
-      case 5:
-        return "Localhost";
       case 1:
-        return "Mainnet";
+        return "Ethereum Mainnet";
+      case 2:
+        return "Cere Mainnet (Testnet)";
       case 3:
-        return "Ropsten";
+        return "Ethereum Ropsten";
       case 4:
-        return "Rinkeby";
-      // case 5:
-      //   return "Goerli";
+        return "Ethereum Rinkeby";
+      case 5:
+        return "Ethereum Goerli";
       case 6:
         return "Kotti";
       case 42:
-        return "Kovan";
+        return "Ethereum Kovan";
       case 61:
         return "Ethereum Classic - Mainnet";
       case 42220:
@@ -69,6 +62,10 @@ export const EVMHomeAdaptorProvider = ({
         return "CELO - Baklava Testnet";
       case 1749641142:
         return "Besu";
+      case 137:
+        return "Polygon Mainnet";
+      case 80001:
+        return "Polygon Mumbai";
       default:
         return "Other";
     }
@@ -81,6 +78,7 @@ export const EVMHomeAdaptorProvider = ({
     handleSetHomeChain,
     homeChains,
     setNetworkId,
+    setWalletType,
   } = useNetworkManager();
 
   const [homeBridge, setHomeBridge] = useState<Bridge | undefined>(undefined);
@@ -118,6 +116,15 @@ export const EVMHomeAdaptorProvider = ({
       onboard
         .walletSelect("metamask")
         .then((success) => {
+          if (window.ethereum) {
+            window.ethereum.on("chainChanged", (ch: any) => {
+              window.location.reload();
+            });
+            window.ethereum.on("accountsChanged", (accounts: any) => {
+              setWalletType("unset");
+            });
+          }
+
           setWalletSelected(success);
           if (success) {
             checkIsReady()
@@ -348,13 +355,16 @@ export const EVMHomeAdaptorProvider = ({
           address,
           (homeChainConfig as EvmBridgeConfig).erc20HandlerAddress
         );
-
+        console.log(
+          "🚀  currentAllowance",
+          utils.formatUnits(currentAllowance, erc20Decimals)
+        );
         if (
           Number(utils.formatUnits(currentAllowance, erc20Decimals)) < amount
         ) {
           if (
             Number(utils.formatUnits(currentAllowance, erc20Decimals)) > 0 &&
-            resetAllowanceLogicFor.includes(tokenAddress)
+            token.isDoubleApproval
           ) {
             //We need to reset the user's allowance to 0 before we give them a new allowance
             //TODO Should we alert the user this is happening here?