+++ src/Modules/TransferActiveModal.tsx
@@ -3,27 +3,30 @@ import React from "react";
 import { makeStyles, createStyles, ITheme } from "@chainsafe/common-theme";
 import {
   Button,
-  ExclamationCircleSvg,
+  NavLink,
   ProgressBar,
   Typography,
 } from "@chainsafe/common-components";
+import ExclamationCircleSvg from "./../media/Icons/exclamation-mark-icon.png";
 import CustomModal from "../Components/Custom/CustomModal";
 import { useChainbridge } from "../Contexts/ChainbridgeContext";
 import { EvmBridgeConfig } from "../chainbridgeConfig";
+import styles from "../Constants/constants";
 
 const useStyles = makeStyles(
   ({ animation, constants, palette, typography }: ITheme) =>
     createStyles({
       root: {
         width: "100%",
+        font: styles.primaryFont,
       },
       inner: {
         width: "100% !important",
         maxWidth: "unset !important",
         display: "flex",
         flexDirection: "row",
-        padding: `${constants.generalUnit * 5}px ${
-          constants.generalUnit * 3.5
+        padding: `${constants.generalUnit * 2}px ${
+          constants.generalUnit * 2
         }px`,
         bottom: 0,
         top: "unset",
@@ -34,20 +37,23 @@ const useStyles = makeStyles(
         transitionDuration: `${animation.transform}ms`,
       },
       heading: {
-        marginBottom: constants.generalUnit,
         whiteSpace: "nowrap",
+        fontSize: 20,
+        marginLeft: constants.generalUnit * 1.5,
+        fontWeight: "bold",
+        fontFamily: styles.primaryFont,
       },
       stepIndicator: {
         ...typography.h4,
         height: 40,
         width: 40,
-        marginRight: constants.generalUnit * 2,
         borderRadius: "50%",
         display: "flex",
         justifyContent: "center",
         alignItems: "center",
-        border: `1px solid ${palette.additional["transactionModal"][2]}`,
-        color: palette.additional["transactionModal"][3],
+        color: "white",
+        background:
+          "linear-gradient(105.79deg, #A700E1 1.84%, #0024E2 102.94%)",
         "& svg": {
           height: 20,
           width: 20,
@@ -57,10 +63,16 @@ const useStyles = makeStyles(
       content: {
         display: "flex",
         flexDirection: "column",
+        width: "100%",
       },
       buttons: {
         display: "flex",
         flexDirection: "row",
+        justifyContent: "center",
+        background:
+          "linear-gradient(105.79deg, #A700E1 1.84%, #0024E2 102.94%)",
+        borderRadius: constants.generalUnit / 2,
+        minHeight: constants.generalUnit * 4,
         marginTop: constants.generalUnit * 5,
         "& > *": {
           textDecoration: "none",
@@ -68,26 +80,42 @@ const useStyles = makeStyles(
         },
       },
       button: {
-        borderColor: `${palette.additional["gray"][8]} !important`,
-        color: `${palette.additional["gray"][8]} !important`,
+        border: "none",
+        color: "white",
+        fontWeight: "bold",
         textDecoration: "none",
         "&:hover": {
-          borderColor: `${palette.additional["gray"][8]} !important`,
-          backgroundColor: `${palette.additional["gray"][8]} !important`,
-          color: `${palette.common.white.main} !important`,
-          textDecoration: "none",
+          background: "none",
+          color: "white",
         },
       },
       initCopy: {
         display: "flex",
         flexDirection: "column",
         justifyContent: "space-between",
+        fontSize: 14,
+        marginLeft: constants.generalUnit * 6.25,
+        color: styles.greyColor,
+        lineHeight: `${constants.generalUnit * 3}px`,
+        fontFamily: styles.secondaryFont,
+        marginBottom: constants.generalUnit * 2,
         "& > *:first-child": {
-          marginTop: constants.generalUnit * 3.5,
-          marginBottom: constants.generalUnit * 8,
+          marginTop: constants.generalUnit * 3,
+          marginBottom: constants.generalUnit * 5,
         },
       },
-      sendingCopy: {},
+      sendingCopy: {
+        color: styles.greyColor,
+        lineHeight: `${constants.generalUnit * 3}px`,
+        fontFamily: styles.secondaryFont,
+        marginLeft: constants.generalUnit * 6.25,
+        marginBottom: constants.generalUnit * 2,
+        fontSize: 14,
+        "& > div:first-child": {
+          fontSize: 14,
+          marginTop: constants.generalUnit * 3,
+        },
+      },
       vote: {
         display: "flex",
         flexDirection: "row",
@@ -111,8 +139,11 @@ const useStyles = makeStyles(
         fontWeight: 600,
       },
       receipt: {
-        marginTop: constants.generalUnit * 3.5,
-        marginBottom: constants.generalUnit * 8,
+        color: styles.greyColor,
+        lineHeight: `${constants.generalUnit * 3}px`,
+        fontFamily: styles.secondaryFont,
+        marginTop: constants.generalUnit * 3,
+        marginLeft: constants.generalUnit * 6.5,
       },
       weighted: {
         fontWeight: 600,
@@ -126,9 +157,37 @@ const useStyles = makeStyles(
           borderRadius: "0 !important",
           "&  >  *": {
             borderRadius: "0 !important",
-            background: `${palette.additional["transactionModal"][1]} !important`,
+            background: `linear-gradient(105.79deg, #A700E1 1.84%, #0024E2 102.94%) !important`,
           },
         },
+      },
+      title: {
+        display: "flex",
+        alignContent: "center",
+        justifyContent: "flex-start",
+        alignItems: "center",
+        marginTop: constants.generalUnit * 3,
+      },
+      exclamation: {
+        height: constants.generalUnit * 3.75,
+        width: constants.generalUnit * 3.75,
+      },
+      footer: {
+        color: styles.primaryTextColor,
+        fontSize: 12,
+        display: "flex",
+        justifyContent: "center",
+        alignItems: "center",
+        paddingTop: constants.generalUnit * 3,
+        fontFamily: styles.secondaryFont,
+        fontStyle: "normal",
+        fontWeight: "bold",
+      },
+      footerText: {
+        color: styles.primaryTextColor,
+        fontSize: 12,
+        fontStyle: "normal",
+        fontWeight: "bold",
       },
     })
 );
@@ -170,41 +229,44 @@ const TransferActiveModal: React.FC<ITransferActiveModalProps> = ({
         variant="primary"
         progress={transactionStatus !== "Transfer Completed" ? -1 : 100}
       />
-      <section>
-        <div className={classes.stepIndicator}>
-          {transactionStatus === "Initializing Transfer" ? (
-            "1"
-          ) : transactionStatus === "In Transit" ? (
-            "2"
-          ) : transactionStatus === "Transfer Completed" ? (
-            "3"
-          ) : (
-            <ExclamationCircleSvg />
-          )}
+      <div className={classes.content}>
+        <div className={classes.title}>
+          <div className={classes.stepIndicator}>
+            {transactionStatus === "Initializing Transfer" ? (
+              "1"
+            ) : transactionStatus === "In Transit" ? (
+              "2"
+            ) : transactionStatus === "Transfer Completed" ? (
+              "3"
+            ) : (
+              <img
+                src={ExclamationCircleSvg}
+                alt="Exclamation"
+                className={classes.exclamation}
+              />
+            )}
+          </div>
+          <Typography className={classes.heading} variant="h3" component="h3">
+            {transactionStatus === "Initializing Transfer"
+              ? "Initializing Transfer"
+              : transactionStatus === "In Transit"
+              ? `In Transit (${
+                  depositVotes < (relayerThreshold || 0)
+                    ? `${depositVotes}/${relayerThreshold} signatures needed`
+                    : "Executing proposal"
+                })`
+              : transactionStatus === "Transfer Completed"
+              ? "Transfer completed"
+              : "Transfer aborted"}
+          </Typography>
         </div>
-      </section>
-      <section className={classes.content}>
-        <Typography className={classes.heading} variant="h3" component="h3">
-          {transactionStatus === "Initializing Transfer"
-            ? "Initializing Transfer"
-            : transactionStatus === "In Transit"
-            ? `In Transit (${
-                depositVotes < (relayerThreshold || 0)
-                  ? `${depositVotes}/${relayerThreshold} signatures needed`
-                  : "Executing proposal"
-              })`
-            : transactionStatus === "Transfer Completed"
-            ? "Transfer completed"
-            : "Transfer aborted"}
-        </Typography>
         {transactionStatus === "Initializing Transfer" ? (
           <div className={classes.initCopy}>
             <Typography>Deposit pending...</Typography>
-            <Typography className={classes.weighted}>
-              This should take a few minutes.
-              <br />
-              Please do not refresh or leave the page.
-            </Typography>
+            <div>
+              This should take a few minutes. <br />
+              <strong> Please do not refresh or leave the page.</strong>
+            </div>
           </div>
         ) : transactionStatus === "In Transit" ? (
           <div className={classes.sendingCopy}>
@@ -230,10 +292,10 @@ const TransferActiveModal: React.FC<ITransferActiveModalProps> = ({
                 );
               }
             })}
-            <Typography className={classes.warning}>
-              This should take a few minutes. <br />
-              Please do not refresh or leave the page.
-            </Typography>
+            <div>
+              This should take a few minutes. <br /> <br />
+              <strong> Please do not refresh or leave the page.</strong>
+            </div>
           </div>
         ) : transactionStatus === "Transfer Completed" ? (
           <>
@@ -304,28 +366,31 @@ const TransferActiveModal: React.FC<ITransferActiveModalProps> = ({
                   View transaction
                 </Button>
               )} */}
-            <section className={classes.buttons}>
-              <Button
-                size="small"
-                className={classes.button}
-                variant="outline"
-                onClick={close}
-              >
-                Start new transfer
-              </Button>
-              <a
-                rel="noopener noreferrer"
-                href={process.env.REACT_APP_SUPPORT_URL}
-                target="_blank"
-              >
-                <Button variant="outline">
-                  Ask a question on {process.env.REACT_APP_SUPPORT_SERVICE}
+            <section>
+              <div className={classes.buttons}>
+                <Button
+                  size="small"
+                  className={classes.button}
+                  variant="outline"
+                  onClick={close}
+                >
+                  Start new transfer
                 </Button>
-              </a>
+              </div>
+              <div className={classes.footer}>
+                <NavLink
+                  style={{ textDecoration: "none" }}
+                  className={classes.footerText}
+                  to={{ pathname: process.env.REACT_APP_SUPPORT_URL }}
+                  target="_blank"
+                >
+                  Ask a question on {process.env.REACT_APP_SUPPORT_SERVICE}
+                </NavLink>
+              </div>
             </section>
           </>
         )}
-      </section>
+      </div>
     </CustomModal>
   );
 };