+++ src/Themes/LightTheme.ts
@@ -46,7 +46,8 @@ export const lightTheme = createTheme({
             root: {
               backgroundColor: "#262626",
               color: "#ffffff",
-              border: `1px solid #262626`,
+              height: 38,
+              borderRadius: 5,
               "& svg": {
                 fill: "#ffffff",
               },
@@ -60,14 +61,14 @@ export const lightTheme = createTheme({
             },
             hover: {
               backgroundColor: "#ffffff",
-              color: "#262626",
+              color: "white",
               "& svg": {
                 fill: "#262626",
               },
             },
             focus: {
               backgroundColor: "#ffffff",
-              color: "#262626",
+              color: "white",
               "& svg": {
                 fill: "#262626",
               },
@@ -77,6 +78,7 @@ export const lightTheme = createTheme({
             root: {
               backgroundColor: "transparent",
               color: "#ffffff",
+              height: 38,
               border: `1px solid #ffffff`,
               "& svg": {
                 fill: "#ffffff",