+++ src/types/globals.ts
@@ -2,6 +2,7 @@ import { ChainbridgeConfig, UIConfig } from "../chainbridgeConfig";
 
 declare global {
   interface Window {
+    ethereum: any;
     __RUNTIME_CONFIG__: {
       UI: UIConfig;
       CHAINBRIDGE: ChainbridgeConfig;