+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - ...
 
+## v0.6.0
+
+- Cherry picked 6159de29, ae130023 from the upstream
+- Turned off double approval
+- Redesign transfer active modal
+- Redesigned transfer and pre-flight transfer page
+
 ## v0.5.0
 
 - Added cross link on header